requirements.txt
openai
PyYAML
paramiko
docker
